+++ framework/src/main/java/com/github/dbmdz/flusswerk/framework/config/FlusswerkConfiguration.java
@@ -99,8 +99,8 @@ public class FlusswerkConfiguration {
   }
 
   @Bean
-  public MessageBroker messageBroker(RoutingProperties routingProperties, RabbitClient rabbitClient)
-      throws IOException {
+  public MessageBroker messageBroker(
+      RoutingProperties routingProperties, RabbitClient rabbitClient) {
     return new MessageBroker(routingProperties, rabbitClient);
   }