+++ framework/src/main/java/com/github/dbmdz/flusswerk/framework/rabbitmq/MessageBroker.java
@@ -4,9 +4,9 @@ import com.github.dbmdz.flusswerk.framework.config.properties.RoutingProperties;
 import com.github.dbmdz.flusswerk.framework.exceptions.InvalidMessageException;
 import com.github.dbmdz.flusswerk.framework.model.Envelope;
 import com.github.dbmdz.flusswerk.framework.model.Message;
-import java.io.IOException;
 import java.util.Collection;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -26,7 +26,7 @@ public class MessageBroker {
 
   private final RabbitClient rabbitClient;
 
-  public MessageBroker(RoutingProperties routing, RabbitClient rabbitClient) throws IOException {
+  public MessageBroker(RoutingProperties routing, RabbitClient rabbitClient) {
     this.routingConfig = routing;
     this.rabbitClient = rabbitClient;
 
@@ -39,32 +39,34 @@ public class MessageBroker {
    * Sends a message to the default output queue as JSON document.
    *
    * @param message the message to send.
-   * @throws IOException if sending the message fails.
    * @deprecated Use {@link Topic#send(Message)} instead
    */
   @Deprecated
-  void send(Message message) throws IOException {
-    var topic = routingConfig.getOutgoing().get("default");
-    if (topic == null) {
+  void send(Message message) {
+    List<String> topics = routingConfig.getOutgoing().get("default");
+    if (topics == null || topics.isEmpty()) {
       throw new RuntimeException("Cannot send message, no default queue specified");
     }
-    send(topic, message);
+    for (String topic : topics) {
+      send(topic, message);
+    }
   }
 
   /**
    * Sends messages to the default output queue as JSON document.
    *
    * @param messages the message to send.
-   * @throws IOException if sending the message fails.
    * @deprecated Use {@link Topic#send(Message)} instead
    */
   @Deprecated
-  public void send(Collection<? extends Message> messages) throws IOException {
-    var topic = routingConfig.getOutgoing().get("default");
-    if (topic == null) {
-      throw new RuntimeException("Cannot send messages, no default queue specified");
+  public void send(Collection<? extends Message> messages) {
+    List<String> topics = routingConfig.getOutgoing().get("default");
+    if (topics == null || topics.isEmpty()) {
+      throw new RuntimeException("Cannot send message, no default queue specified");
     }
-    send(topic, messages);
+    for (String topic : topics) {
+      send(topic, messages);
+    }
   }
 
   /**
@@ -73,9 +75,8 @@ public class MessageBroker {
    * @param routingKey the routing key for the queue to send the message to (usually the queue
    *     name).
    * @param message the message to send.
-   * @throws IOException if sending the message fails.
    */
-  void send(String routingKey, Message message) throws IOException {
+  void send(String routingKey, Message message) {
     rabbitClient.send(routingConfig.getExchange(routingKey), routingKey, message);
   }
 
@@ -90,9 +91,8 @@ public class MessageBroker {
    * @param routingKey the routing key for the queue to send the message to (usually the queue
    *     name).
    * @param messages the messages to send.
-   * @throws IOException if sending a message fails.
    */
-  void send(String routingKey, Collection<? extends Message> messages) throws IOException {
+  void send(String routingKey, Collection<? extends Message> messages) {
     for (Message message : messages) {
       send(routingKey, message);
     }
@@ -159,7 +159,7 @@ public class MessageBroker {
     }
   }
 
-  private void provideInputQueues() throws IOException {
+  private void provideInputQueues() {
     for (String inputQueue : routingConfig.getIncoming()) {
       FailurePolicy failurePolicy = routingConfig.getFailurePolicy(inputQueue);
       final String exchange = routingConfig.getExchange(inputQueue);
@@ -187,8 +187,8 @@ public class MessageBroker {
     }
   }
 
-  private void provideOutputQueues() throws IOException {
-    for (String topic : routingConfig.getOutgoing().values()) {
+  private void provideOutputQueues() {
+    for (String topic : routingConfig.allOutgoing()) {
       rabbitClient.declareQueue(
           topic,
           routingConfig.getExchange(topic),
@@ -212,9 +212,8 @@ public class MessageBroker {
    *
    * @param message the message to reject
    * @return true if retry, false if failed
-   * @throws IOException if communication with RabbitMQ failed
    */
-  public boolean reject(Message message) throws IOException {
+  public boolean reject(Message message) {
     final Envelope envelope = message.getEnvelope();
     final long maxRetries = routingConfig.getFailurePolicy(message).getRetries();
     if (envelope.getRetries() < maxRetries) {
@@ -227,7 +226,7 @@ public class MessageBroker {
     }
   }
 
-  void fail(Message message, boolean ackMessage) throws IOException {
+  void fail(Message message, boolean ackMessage) {
     if (ackMessage) {
       ack(message);
     }
@@ -239,11 +238,11 @@ public class MessageBroker {
     }
   }
 
-  public void fail(Message message) throws IOException {
+  public void fail(Message message) {
     fail(message, true);
   }
 
-  public void retry(Message message) throws IOException {
+  public void retry(Message message) {
     LOGGER.debug("Send message to retry queue: " + message);
     FailurePolicy failurePolicy = routingConfig.getFailurePolicy(message);
     String retryRoutingKey = failurePolicy.getRetryRoutingKey();
@@ -254,7 +253,7 @@ public class MessageBroker {
     }
   }
 
-  private void provideExchanges() throws IOException {
+  private void provideExchanges() {
     for (String exchange : routingConfig.getExchanges()) {
       rabbitClient.provideExchange(exchange);
     }
@@ -267,11 +266,10 @@ public class MessageBroker {
    * Returns the number of messages in known queues
    *
    * @return a map of queue names and the number of messages in these queues
-   * @throws IOException if communication with RabbitMQ fails
    * @deprecated Use {@link Queue#messageCount()} instead.
    */
   @Deprecated
-  Map<String, Long> getMessageCounts() throws IOException {
+  Map<String, Long> getMessageCounts() {
     Map<String, Long> result = new HashMap<>();
     for (String queue : routingConfig.getIncoming()) {
       result.put(queue, rabbitClient.getMessageCount(queue));
@@ -279,7 +277,7 @@ public class MessageBroker {
     return result;
   }
 
-  Map<String, Long> getFailedMessageCounts() throws IOException {
+  Map<String, Long> getFailedMessageCounts() {
     Map<String, Long> result = new HashMap<>();
     for (String inputQueue : routingConfig.getIncoming()) {
       FailurePolicy failurePolicy = routingConfig.getFailurePolicy(inputQueue);
@@ -291,7 +289,7 @@ public class MessageBroker {
     return result;
   }
 
-  public Map<String, Long> getRetryMessageCounts() throws IOException {
+  public Map<String, Long> getRetryMessageCounts() {
     Map<String, Long> result = new HashMap<>();
     for (String inputQueue : routingConfig.getIncoming()) {
       FailurePolicy failurePolicy = routingConfig.getFailurePolicy(inputQueue);