+++ framework/src/main/java/com/github/dbmdz/flusswerk/framework/rabbitmq/Topic.java
@@ -4,7 +4,6 @@ import static java.util.Objects.requireNonNull;
 
 import com.github.dbmdz.flusswerk.framework.model.Message;
 import com.github.dbmdz.flusswerk.framework.reporting.Tracing;
-import java.io.IOException;
 import java.util.Collection;
 import java.util.List;
 import java.util.Objects;
@@ -13,7 +12,7 @@ import java.util.Objects;
  * Represents a AMQP/RabbitMQ topic to send messages to. In many setups this is equal to the
  * respective queue name.
  */
-public class Topic {
+public class Topic implements Sender {
 
   private final String name;
   private final MessageBroker messageBroker;
@@ -31,10 +30,9 @@ public class Topic {
    * case, every time you call this method it creates a new tracing path.
    *
    * @param message The message to send.
-   * @throws IOException If communication with RabbitMQ fails or if the message cannot be serialized
-   *     to JSON.
    */
-  public void send(Message message) throws IOException {
+  @Override
+  public void send(Message message) {
     // Only set a tracing path if there is none yet
     if (message.getTracing() == null || message.getTracing().isEmpty()) {
       message.setTracing(getTracingPath());
@@ -46,10 +44,9 @@ public class Topic {
    * Sends multiple messages to this topic.
    *
    * @param messages The messages to send.
-   * @throws IOException If communication with RabbitMQ fails or if the message cannot be serialized
-   *     to JSON.
    */
-  public void send(Collection<Message> messages) throws IOException {
+  @Override
+  public void send(Collection<Message> messages) {
     // Get a new tracing path in case one is needed
     final List<String> tracingPath = getTracingPath();
     messages.stream()
@@ -62,10 +59,9 @@ public class Topic {
    * Convenience implementation, mostly for tests.
    *
    * @param messages The messages to send.
-   * @throws IOException If communication with RabbitMQ fails or if the message cannot be serialized
-   *     to JSON.
    */
-  public void send(Message... messages) throws IOException {
+  @Override
+  public void send(Message... messages) {
     send(List.of(messages));
   }
 
@@ -76,6 +72,7 @@ public class Topic {
    *
    * @param message The message serialized to bytes
    */
+  @Override
   public void sendRaw(byte[] message) {
     messageBroker.sendRaw(name, message);
   }